Refactor to help with packaging
