+++ src/raster_scripts/shapefile_manipulation.py
